+++ plugins/Institution/src/Controller/InstitutionsController.php
@@ -10324,22 +10324,51 @@ class InstitutionsController extends AppController
         if (empty($id)) {
             throw new NotFoundException(__('Invalid file'));
         }
+         $url = $_SERVER['HTTP_REFERER'] ?? '';
+        if (strpos($url, '/StudentEnrolment') !== false) {
+            $studentEnrolmentCustomFieldValues = TableRegistry::getTableLocator()->get('StudentCustomField.StudentCustomFieldValues');
+            $getQueryString = $this->getQueryString();
+            $fileRecord = $studentEnrolmentCustomFieldValues->find()
+                            ->where([
+                                'file IS NOT' => null,
+                                'id' => $getQueryString['id']
+                            ])->first();
 
+            if (empty($fileRecord) || empty($fileRecord->file)) {
+                throw new NotFoundException(__('File not found'));
+            }
+            $fileName = $fileRecord->text_value;
+            $fileResource = $fileRecord->file;
+            
+        }elseif(strpos($url, '/StudentAdmission') !== false) {
+            $studentAdmissionCustomFieldValues = TableRegistry::getTableLocator()->get('StudentCustomField.StudentAdmissionCustomFieldValues');
+            $getQueryString = $this->getQueryString();
+            $fileRecord = $studentAdmissionCustomFieldValues->find()
+                            ->where([
+                                'file IS NOT' => null,
+                                'id' => $getQueryString['id']
+                            ])->first();
+
+            if (empty($fileRecord) || empty($fileRecord->file)) {
+                throw new NotFoundException(__('File not found'));
+            }
+            $fileName = $fileRecord->text_value;
+            $fileResource = $fileRecord->file;
+        }else{
         // Load your custom field values table
-        $InstitutionCustomFieldValues = TableRegistry::getTableLocator()->get('InstitutionCustomField.InstitutionCustomFieldValues');
-        $fileRecord = $InstitutionCustomFieldValues->find()
-                        ->where([
-                            'file IS NOT' => null,
-                            'file_name IS NOT' => null,
-                            'institution_id' => $this->getInstitutionID(),
-                        ])->first();
+            $InstitutionCustomFieldValues = TableRegistry::getTableLocator()->get('InstitutionCustomField.InstitutionCustomFieldValues');
+            $fileRecord = $InstitutionCustomFieldValues->find()
+                            ->where([
+                                'file IS NOT' => null,
+                                'institution_id IS' => $this->getInstitutionID(),
+                            ])->first();
 
-        if (empty($fileRecord) || empty($fileRecord->file_name) || empty($fileRecord->file)) {
-            throw new NotFoundException(__('File not found'));
+            if (empty($fileRecord) || empty($fileRecord->file)) {
+                throw new NotFoundException(__('File not found'));
+            }
+            $fileName = $fileRecord->text_value;
+            $fileResource = $fileRecord->file;
         }
-
-        $fileName = $fileRecord->file_name;
-        $fileResource = $fileRecord->file;
         $this->response = $this->response
             ->withType(mime_content_type($fileResource))
             ->withHeader('Content-Disposition', 'attachment; filename="' . $fileName . '"')
@@ -10467,6 +10496,3 @@ class InstitutionsController extends AppController
     }
 
 }
-
-
-