+++ webroot/CHANGELOG.md
@@ -1,4 +1,6 @@
 ### 5.17.0 - 2026-08-19
+- Institution Bugs: Gives 404 on addition and deletion, however performs that action.
+- Bug Fixed: Registrations: When registering student with custom field file and file is uploaded, file is not seen in student list page.
 - Bug Fixed: Remove unneccessary files on filesystem
 - Bug Fixed: Docker container not working.
 - Bug Fixed: Core>Directory>Search user records>Merge : Investigate why merge failed